requirements.txt
streamlit
numpy
matplotlib
scikit-learn
scikit-image
Pillow
transformers
torch
